fix: code review feedback
