+++ src/app/manifest.ts
@@ -14,12 +14,6 @@ export default function manifest(): MetadataRoute.Manifest {
     scope: "/",
     icons: [
       {
-        src: "/apple-touch-icon.png",
-        sizes: "180x180",
-        type: "image/png",
-        purpose: "maskable",
-      },
-      {
         src: "/web-app-manifest-192x192.png",
         sizes: "192x192",
         type: "image/png",
