+++ src/app/sign-in/_components/sign-in-redirect.tsx
@@ -7,15 +7,20 @@ import { useEffect } from "react";
 
 export function SignInRedirect() {
   const searchParams = useSearchParams();
-  const redirectUrl = searchParams.get("redirect_url") || ROUTES.DASHBOARD;
-  const auth = useConvexAuth();
+  const rawRedirect = searchParams.get("redirect_url");
+  // Only allow same-origin relative paths like "/foo" (reject "//" and absolute URLs)
+  const redirectUrl =
+    rawRedirect && /^\/(?!\/)/.test(rawRedirect)
+      ? rawRedirect
+      : ROUTES.DASHBOARD;
+  const { isAuthenticated } = useConvexAuth();
   const router = useRouter();
 
   useEffect(() => {
-    if (auth.isAuthenticated) {
-      router.push(redirectUrl);
+    if (isAuthenticated) {
+      router.replace(redirectUrl);
     }
-  }, [auth, redirectUrl, router]);
+  }, [isAuthenticated, redirectUrl, router]);
 
   return null;
 }
