+++ src/components/workout-drawer/index.ts
@@ -2,6 +2,7 @@ export { CreateSplitDialog } from "./create-split-dialog";
 export { ExerciseFilter } from "./exercise-filter";
 export { ExerciseSetForm } from "./exercise-set-form";
 export { SelectExerciseDrawer } from "./select-exercise-drawer";
+export { SplitListItem } from "./split-list-item";
 export { SplitSelectionGrid } from "./split-selection-grid";
 export { SplitSelectionModal } from "./split-selection-modal";
 export { WeightUnitToggle } from "./weight-unit-toggle";
